backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
sqlalchemy>=2.0.25
python-multipart>=0.0.9
python-dotenv>=1.0.1